# CLN/PERF: no need for kahan for int group_cumsum